fresh screens; templating for content_scripts
