+++ source/chrome/manifest.json
@@ -1,14 +1,14 @@
 {
 	"manifest_version": 2,
 	"name": "Some Rich Asshole Enhanced",
-	"version": "3.0.6",
+	"version": "3.0.7",
 	"description": "Replaces He-Who-Must-Not-Be-Named's name, and those of some related individuals, with more colorful descriptors.",
   "options_page": "options.html",
   "permissions": [
     "storage"
   ],
   "background": {
-    "scripts": ["jquery-3.1.1.min.js", "background.js"],
+    "scripts": ["background.js"],
     "persistent": false
   },
 	"content_scripts":
